move to package extensions, no other changes
