+++ Project.toml
@@ -1,23 +1,32 @@
 name = "TransmuteDims"
 uuid = "24ddb15e-299a-5cc3-8414-dbddc482d9ca"
 authors = ["Michael Abbott"]
-version = "0.1.17"
+version = "0.1.18"
 
 [deps]
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+
+[weakdeps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 GPUArraysCore = "46192b85-c4d5-4398-a991-12ede77f4527"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
+OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Strided = "5e0ebb24-38b0-5f93-81fe-25c709ecae67"
+Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
+
+[extensions]
+TD_ChainRulesCore = "ChainRulesCore"
+TD_GPU = ["Adapt", "GPUArraysCore"]
+TD_OffsetArrays = "OffsetArrays"
+TD_Strided = "Strided"
+TD_Tracker = "Tracker"
 
 [compat]
 Adapt = "3.3, 4"
 BenchmarkTools = "1.2"
 ChainRulesCore = "1.11"
-Documenter = "0.27"
+Documenter = "1.10"
 GPUArraysCore = "0.1.2, 0.2"
-Requires = "1"
 StaticArrays = "1"
 Strided = "2"
 Tracker = "0.2.16"
@@ -28,12 +37,10 @@ julia = "1.10"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 JLArrays = "27aeb0d3-9eb9-45fb-866b-73c2ecf80fcb"
-OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
-test = ["Test", "BenchmarkTools", "Documenter", "JLArrays", "OffsetArrays", "Random", "StaticArrays", "Tracker", "Zygote"]
+test = ["Test", "BenchmarkTools", "Documenter", "JLArrays", "OffsetArrays", "Random", "Strided", "StaticArrays", "Tracker", "Zygote"]
